requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.6.5
amex-guardrails==0.0.4
amex-nemoguardrails==0.0.7
amexcerts==0.1.2
annoy==1.17.3
anyio==3.7.1
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
backoff==2.2.1
bcrypt==3.2.0
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
cryptography==41.0.7
dataclasses-json==0.6.4
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
distro==1.9.0
dnspython==2.7.0
email-validator==1.1.3
executing==2.0.1
fastapi==0.96.0
Flask==3.0.3
frozenlist==1.4.1
fstr==0.1.0a2
googleapis-common-protos==1.62.0
grpcio==1.62.0
gunicorn==23.0.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.6
ipykernel==6.29.4
ipython==8.24.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
langchain==0.1.13
langchain-community==0.0.31
langchain-core==0.1.46
langchain-openai==0.1.3
langchain-text-splitters==0.0.1
langsmith==0.1.51
Mako==1.3.9
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib-inline==0.1.7
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.5.6
numpy==1.26.4
openai==1.23.6
opentelemetry-api==1.12.0
opentelemetry-exporter-otlp==1.12.0
opentelemetry-exporter-otlp-proto-grpc==1.12.0
opentelemetry-exporter-otlp-proto-http==1.12.0
opentelemetry-instrumentation==0.41b0
opentelemetry-instrumentation-logging==0.41b0
opentelemetry-proto==1.12.0
opentelemetry-sdk==1.12.0
opentelemetry-semantic-conventions==0.33b0
orjson==3.10.1
packaging==23.2
parso==0.8.3
passlib==1.7.4
pexpect==4.9.0
platformdirs==4.2.1
prompt-toolkit==3.0.43
protobuf==3.20.2
psutil==5.9.8
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.14
Pygments==2.17.2
PyJWT==2.10.1
pyOpenSSL==23.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-editor==1.0.4
python-multipart==0.0.20
PyYAML==6.0.1
pyzmq==26.0.2
regex==2023.12.25
requests==2.31.0
rfc3986==1.5.0
safechain==0.1.0
simpleeval==0.9.13
six==1.12.0
sniffio==1.3.1
SQLAlchemy @ file:///Users/smish257/Downloads/SQLAlchemy-1.4.54-cp311-cp311-macosx_10_9_universal2.whl#sha256=b05e0626ec1c391432eabb47a8abd3bf199fb74bfde7cc44a26d2b1b352c2c6e
stack-data==0.6.3
starlette==0.27.0
tenacity==8.2.3
tiktoken==0.6.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.22.0
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
